+++ src/pymadng/madp_classes.py
@@ -2,26 +2,26 @@ import warnings # To warn the user when they try to deepcopy a mad_ref
 from typing import Iterable, Union, Any  # To make stuff look nicer
 import numpy as np
 from .madp_pymad import mad_process, mad_ref, type_str, is_private
-from .madp_strings import get_args_string, get_kwargs_string
+from .madp_strings import format_args_to_string, format_kwargs_to_string
 
-# TODO: Are you able to store the actual parent?
-# TODO: Allow __setitem__ to work with multiple indices (Should be a simple recursive loop)
+# TODO: Are you able to store the actual parent? (jgray 2023)
+# TODO: Allow __setitem__ to work with multiple indices (Should be a simple recursive loop) (jgray 2023)
 
 MADX_methods = ["load", "open_env", "close_env"]
 
-
+# MAD High Level reference
 class madhl_ref(mad_ref):
   def __init__(self, name: str, mad_proc: mad_process):
     super(madhl_ref, self).__init__(name, mad_proc)
     self._parent = (
       "[" in name and "[".join(name.split("[")[:-1]) or None
     )  # if name is compound, get parent by string manipulation
-    self._lst_cntr = mad_proc.lst_cntr
+    self._lst_cntr = mad_proc.lst_cntr # Set the last counter to the current value in the process
 
   def __setattr__(self, item, value):
-    if is_private(item):
+    if is_private(item): # If the attribute is private, set it as a normal attribute
       return super(madhl_ref, self).__setattr__(item, value)
-    self[item] = value
+    self[item] = value # Otherwise, set the item as a variable in the MAD-NG process
 
   def __setitem__(
     self,
@@ -29,47 +29,48 @@ class madhl_ref(mad_ref):
     value: Union[str, int, float, np.ndarray, bool, list],
   ):
     if isinstance(item, int):
-      self._mad.send_vars(**{f"{self._name}[{item+1}]": value})
+      item = item + 1 # Ints need to be incremented by 1 to match MAD-NG indexing
     elif isinstance(item, str):
-      self._mad.send_vars(**{f"{self._name}['{item}']": value})
-    else:
+      item = f"'{item}'" # Strings need to be wrapped in quotes
+    else: # Any other index type is invalid
       raise TypeError(
         "Cannot index type of ", type(item), "expected string or int"
       )
+    self._mad.send_vars(**{f"{self._name}[{item+1}]": value}) 
 
   def __add__(self, rhs):
-    return self.__gOp__(rhs, "+")
+    return self.__generate_operation__(rhs, "+")
 
   def __mul__(self, rhs):
-    return self.__gOp__(rhs, "*")
+    return self.__generate_operation__(rhs, "*")
 
   def __pow__(self, rhs):
-    return self.__gOp__(rhs, "^")
+    return self.__generate_operation__(rhs, "^")
 
   def __sub__(self, rhs):
-    return self.__gOp__(rhs, "-")
+    return self.__generate_operation__(rhs, "-")
 
   def __truediv__(self, rhs):
-    return self.__gOp__(rhs, "/")
+    return self.__generate_operation__(rhs, "/")
 
   def __mod__(self, rhs):
-    return self.__gOp__(rhs, "%")
+    return self.__generate_operation__(rhs, "%")
 
   def __eq__(self, rhs):
     if isinstance(rhs, type(self)) and self._name == rhs._name:
       return True
     else:
-      return self.__gOp__(rhs, "==").eval()
+      return self.__generate_operation__(rhs, "==").eval()
 
-  def __gOp__(self, rhs, operator: str):
+  def __generate_operation__(self, rhs, operator: str):
     rtrn = madhl_reflast(self._mad)
-    self._mad.psend(
+    self._mad.protected_send(
       f"{rtrn._name} = {self._name} {operator} {self._mad.py_name}:recv()"
     ).send(rhs)
     return rtrn
 
   def __len__(self):
-    return self._mad.precv(f"#{self._name}")
+    return self._mad.safe_recv_var(f"#{self._name}")
 
   def __str__(self):
     val = self._mad.recv_vars(self._name)
@@ -81,14 +82,14 @@ class madhl_ref(mad_ref):
   def eval(self):
     return self._mad.recv_vars(self._name)
 
-  def __repr__(self):
-    return f"MAD-NG Object(Name: {self._name}, Parent: {self._parent})"
+  def __repr__(self): # TODO: This should be better (jgray 2024)
+    return f"MAD-NG Object(Name: {self._name}, Parent: {self._parent}, Process: {repr(self._mad)})"
 
   def __dir__(self) -> Iterable[str]:
     name = self._name
     if name[:5] == "_last":
       name = name + ".__metatable or " + name
-    self._mad.psend(f"""
+    self._mad.protected_send(f"""
     local modList={{}}; local i = 1;
     for modname, mod in pairs({name}) do modList[i] = modname; i = i + 1; end
     {self._mad.py_name}:send(modList)
@@ -110,10 +111,10 @@ class madhl_ref(mad_ref):
 class madhl_obj(madhl_ref):
   def __dir__(self) -> Iterable[str]:
     if not self._mad.ipython_use_jedi:
-      self._mad.psend(
+      self._mad.protected_send(
         f"{self._mad.py_name}:send({self._name}:get_varkeys(MAD.object))"
       )
-    varnames = self._mad.precv(f"{self._name}:get_varkeys(MAD.object, false)")
+    varnames = self._mad.safe_recv_var(f"{self._name}:get_varkeys(MAD.object, false)")
 
     if not self._mad.ipython_use_jedi:
       varnames.extend([x + "()" for x in self._mad.recv() if not x in varnames])
@@ -121,10 +122,10 @@ class madhl_obj(madhl_ref):
 
   def __call__(self, *args, **kwargs):
     last_obj = madhl_objlast(self._mad)
-    kwargs_str, kwargs_to_send = get_kwargs_string(self._mad.py_name, **kwargs)
-    args_str, args_to_send = get_args_string(self._mad.py_name, *args)
+    kwargs_str, kwargs_to_send = format_kwargs_to_string(self._mad.py_name, **kwargs)
+    args_str, args_to_send = format_args_to_string(self._mad.py_name, *args)
 
-    self._mad.psend(
+    self._mad.protected_send(
       f"{last_obj._name} = __mklast__( {self._name} {{ {kwargs_str[1:-1]} {args_str} }} )"
     )
     for var in kwargs_to_send + args_to_send:
@@ -132,13 +133,13 @@ class madhl_obj(madhl_ref):
     return last_obj
 
   def __iter__(self):
-    self.__iterIndex__ = -1
+    self._iterindex = -1
     return self
 
   def __next__(self):
     try:
-      self.__iterIndex__ += 1
-      return self[self.__iterIndex__]
+      self._iterindex += 1
+      return self[self._iterindex]
     except IndexError:
       raise StopIteration
 
@@ -153,23 +154,24 @@ class madhl_obj(madhl_ref):
     Returns:
             pandas.DataFrame or tfs.TfsDataFrame: The dataframe containing the object's data.
     """
-    if not self._mad.precv(f"MAD.typeid.is_mtable({self._name})"):
+    if not self._mad.safe_recv_var(f"MAD.typeid.is_mtable({self._name})"):
       raise TypeError("Object is not a table, cannot convert to dataframe")
 
     import pandas as pd
 
     try:
       import tfs
-
-      DataFrame, hattr = tfs.TfsDataFrame, "headers"
-    except ImportError:
-      DataFrame, hattr = pd.DataFrame, "attrs"
+      # If tfs is available, use the headers attribute
+      DataFrame, hdr_attr = tfs.TfsDataFrame, "headers"
+    except ImportError: 
+      # Otherwise, use the pandas dataframe and attrs attribute
+      DataFrame, hdr_attr = pd.DataFrame, "attrs"
 
     py_name, obj_name = self._mad.py_name, self._name
-    self._mad.psend( # Sending every value individually is slow (sending vectors is fast)
+    self._mad.protected_send( # Sending every value individually is slow (sending vectors is fast)
       f"""
-local is_vector, is_number in MAD.typeid
-local colnames = {obj_name}:colnames() -- Get the column names 
+local is_vector, is_number, is_string in MAD.typeid
+local colnames = {py_name}:recv() or {obj_name}:colnames() -- Get the column names 
 {py_name}:send(colnames)               -- Send the column names
 
 -- Loop through all the column names and send them with their data
@@ -177,13 +179,22 @@ for i, colname in ipairs(colnames) do
   local col = {obj_name}:getcol(colname)
 
   -- If the column is not a vector and has a metatable, then convert it to a table (reference or generator columns)
-  if not is_vector(col) and getmetatable(col) then
-    local tbl, conv_to_vec = table.new(#col, 0), true
+  if not is_vector(col) or getmetatable(col) then
+    local tbl = table.new(#col, 0)
+    local conv_to_vec = true
+    local conv_to_str = true
     for i, val in ipairs(col) do 
     -- From testing, checking if I can convert to a vector is faster than sending the table
-      conv_to_vec, tbl[i] = conv_to_vec and is_number(val), val
+      conv_to_vec = conv_to_vec and is_number(val)
+      conv_to_str = conv_to_str and is_string(val)
+      tbl[i] = val
     end
-    col = conv_to_vec and MAD.vector(tbl) or tbl
+    if conv_to_str then
+      tbl = table.concat(tbl, "\\n")
+    elseif conv_to_vec then
+      tbl = MAD.vector(tbl)
+    end
+    col = tbl
   end
 
   {py_name}:send(col) -- Send the column data
@@ -197,6 +208,7 @@ for i, attr in ipairs(header) do
 end
 """
     )
+    self._mad.send(columns)
     # Create the dataframe from the data sent
     colnames = self._mad.recv()
     full_tbl = { # The string is in case references are within the table
@@ -204,19 +216,19 @@ end
     }
 
     # Get the header names and data
-    hnams = self._mad.recv()
-    header = {hnam: self._mad.recv(f"{obj_name}['{hnam}']") for hnam in hnams}
+    hdr_names = self._mad.recv()
+    hdr = {hdr_name: self._mad.recv(f"{obj_name}['{hdr_name}']") for hdr_name in hdr_names}
     
     # Not keen on the .squeeze() but it works (ng always sends 2D arrays, but I need the columns in 1D)
     for key, val in full_tbl.items():
       if isinstance(val, np.ndarray):
         full_tbl[key] = val.squeeze()
+      elif isinstance(val, str):
+        full_tbl[key] = val.split("\n")
       
     # Now create the dataframe
     df = DataFrame(full_tbl)
-    if columns:
-      df = df[columns] # Only keep the columns specified
-    setattr(df, hattr, header)
+    setattr(df, hdr_attr, hdr)
 
     return df
 
@@ -226,8 +238,8 @@ class madhl_fun(madhl_ref):
   def __call_func(self, funcName: str, *args):
     """Call the function funcName and store the result in ``_last``."""
     rtrn_ref = madhl_reflast(self._mad)
-    args_string, vars_to_send = get_args_string(self._mad.py_name, *args)
-    self._mad.psend(f"{rtrn_ref._name} = __mklast__({funcName}({args_string}))\n")
+    args_string, vars_to_send = format_args_to_string(self._mad.py_name, *args)
+    self._mad.protected_send(f"{rtrn_ref._name} = __mklast__({funcName}({args_string}))\n")
     for var in vars_to_send:
       self._mad.send(var)
     return rtrn_ref
@@ -238,18 +250,20 @@ class madhl_fun(madhl_ref):
     # Checks for MADX methods
     call_from_madx = (
       self._parent and self._parent.split("['")[-1].strip("']") == "MADX"
-    )
-    if call_from_madx:
+    ) 
+    if call_from_madx: # Retrive the function name if the direct parent is MADX
       funcname = self._name.split("['")[-1].strip("']")
 
     ismethod = self._parent and (
-      self._mad.precv(
+      self._mad.safe_recv_var(
         f"""
   MAD.typeid.is_object({self._parent}) or MAD.typeid.isy_matrix({self._parent})
   """
       )
-    )
-    if ismethod and not (call_from_madx and not funcname in MADX_methods):
+    ) # Identify if _parent needs to be sent as the first argument (for methods)
+
+    # If it is a method, and if it is a MADX method when called from MADX
+    if ismethod and not (call_from_madx and not funcname in MADX_methods): 
       return self.__call_func(self._name, self._parent, *args)
     else:
       return self.__call_func(self._name, *args)