+++ src/pymadng/madp_pymad.py
@@ -1,6 +1,12 @@
-import struct, os, subprocess, sys, select
+import struct
+import os
+import subprocess 
+import sys
+import select
+import signal
 from typing import Union, Callable, Any
 import numpy as np
+from pathlib import Path
 
 __all__ = ["mad_process"]
 
@@ -13,31 +19,52 @@ def is_private(varname):
 
 
 class mad_process:
-  def __init__(self, mad_path: str, py_name: str = "py", debug: bool = False) -> None:
+  def __init__(self, mad_path: Union[str, Path], py_name: str = "py", debug: Union[int, str, bool] = False) -> None:
     self.py_name = py_name
 
+    mad_path = Path(mad_path)
+    if not mad_path.exists():
+      raise FileNotFoundError(f"Could not find MAD executable at {mad_path}")
+
     # Create the pipes for communication
-    self.from_mad, mad_write = os.pipe()
-    mad_read, self.to_mad = os.pipe()
+    self.mad_output_pipe, mad_write = os.pipe()
+    mad_read, self.mad_input_pipe = os.pipe()
 
     # Open the pipes for communication to MAD (the stdin of MAD)
-    self.fto_mad = os.fdopen(self.to_mad, "wb", buffering=0)
+    self.mad_input_stream = os.fdopen(self.mad_input_pipe, "wb", buffering=0)
+
+    if isinstance(debug, str):
+      debug_file = open(debug, "w")
+      stdout = debug_file.fileno()
+    elif isinstance(debug, bool):
+      stdout = sys.stdout.fileno()
+    elif isinstance(debug, int):
+      stdout = debug
+    else:
+      raise TypeError("Debug must be a file name, file descriptor or a boolean")
 
     # Create a chunk of code to start the process
     startupChunk = (
-      f"MAD.pymad '{py_name}' {{_dbg = {str(debug).lower()}}} :__ini({mad_write})"
+      f"MAD.pymad '{py_name}' {{_dbg = {str(bool(debug)).lower()}}} :__ini({mad_write})"
     )
+    original_sigint_handler = signal.getsignal(signal.SIGINT)
+    def delete_process(sig, frame):
+        self.process.terminate()  # In case user left mad waiting
+        self.mad_input_stream.close()
+        self.process.wait()
+        signal.signal(signal.SIGINT, original_sigint_handler)
+        raise KeyboardInterrupt("MAD process was interrupted, and has been deleted")
+    signal.signal(signal.SIGINT, delete_process)  # Delete the process if interrupted
 
     # Start the process
     self.process = subprocess.Popen(
-      [mad_path, "-q", "-e", startupChunk],
+      [str(mad_path), "-q", "-e", startupChunk],
       bufsize=0,
       stdin=mad_read,  # Set the stdin of MAD to the read end of the pipe
-      stdout=sys.stdout.fileno(),  # Forward stdout
-      preexec_fn=os.setpgrp,  # Don't forward signals
+      stdout=stdout,  # Forward stdout
       pass_fds=[
         mad_write,
-        sys.stdout.fileno(),
+        stdout,
         sys.stderr.fileno(),
       ],  # Don't close these (python closes all fds by default)
     )
@@ -47,48 +74,50 @@ class mad_process:
     os.close(mad_read)
 
     # Create a global variable dictionary for the exec function (could be extended to include more variables)
-    self.globalVars = {"np": np}
+    self.python_exec_context = {"np": np}
 
     # Open the pipe from MAD (this is where MAD will no longer hang)
-    self.ffrom_mad = os.fdopen(self.from_mad, "rb")
+    self.mad_read_stream = os.fdopen(self.mad_output_pipe, "rb")
+    self.history = "" # Begin the recording of the history
 
     # stdout should be line buffered by default, but for jupyter notebook,
     # stdout is redirected and not line buffered by default
-    self.send(
-      f"""io.stdout:setvbuf('line')
-    {self.py_name}:send(1)"""
-    )
+    self.send("io.stdout:setvbuf('line')")
+    
+    # Check if MAD started successfully
+    self.send(f"{self.py_name}:send(1)")
+    startup_status_checker = select.select([self.mad_read_stream], [], [], 10)  # May not work on windows
 
     # Check if MAD started successfully using select
-    checker = select.select([self.ffrom_mad], [], [], 1)  # May not work on windows
-    if not checker[0] or self.recv() != 1:  # Need to check number?
+    if not startup_status_checker[0] or self.recv() != 1:  # Need to check number?
+      del self
       raise OSError(f"Unsuccessful starting of {mad_path} process")
 
-  def send_rng(self, start: float, stop: float, size: int):
+  def send_range(self, start: float, stop: float, size: int):
     """Send a numpy array as a rng to MAD"""
-    self.fto_mad.write(b"rng_")
-    send_grng(self, start, stop, size)
+    self.mad_input_stream.write(b"rng_")
+    send_generic_range(self, start, stop, size)
 
-  def send_lrng(self, start: float, stop: float, size: int):
+  def send_logrange(self, start: float, stop: float, size: int):
     """Send a numpy array as a logrange to MAD"""
-    self.fto_mad.write(b"lrng")
-    send_grng(self, start, stop, size)
+    self.mad_input_stream.write(b"lrng")
+    send_generic_range(self, start, stop, size)
 
   def send_tpsa(self, monos: np.ndarray, coefficients: np.ndarray):
     """Send the monomials and coeeficients of a TPSA to MAD, creating a table representing the TPSA object"""
-    self.fto_mad.write(b"tpsa")
-    send_gtpsa(self, monos, coefficients, send_num)
+    self.mad_input_stream.write(b"tpsa")
+    send_generic_tpsa(self, monos, coefficients, send_num)
 
-  def send_ctpsa(self, monos: np.ndarray, coefficients: np.ndarray):
+  def send_cpx_tpsa(self, monos: np.ndarray, coefficients: np.ndarray):
     """Send the monomials and coeeficients of a complex TPSA to MAD, creating a table representing the complex TPSA object"""
-    self.fto_mad.write(b"ctpa")
-    send_gtpsa(self, monos, coefficients, send_cpx)
+    self.mad_input_stream.write(b"ctpa")
+    send_generic_tpsa(self, monos, coefficients, send_cpx)
 
   def send(self, data: Union[str, int, float, np.ndarray, bool, list]):
     """Send data to MAD, returns self for chaining"""
     try:
       typ = type_str[get_typestr(data)]
-      self.fto_mad.write(typ.encode("utf-8"))
+      self.mad_input_stream.write(typ.encode("utf-8"))
       type_fun[typ]["send"](self, data)
       return self
     except KeyError:  # raise not in exception to reduce error output
@@ -96,21 +125,21 @@ class mad_process:
         f"Unsupported data type, expected a type in: \n{list(type_str.keys())}, got {type(data)}"
       ) from None
 
-  def psend(self, string: str):
+  def protected_send(self, string: str):
     """Perform a protected send to MAD, by first enabling error handling, so that if an error occurs, an error is returned"""
     return self.send(f"{self.py_name}:__err(true); {string}; {self.py_name}:__err(false);")
 
-  def precv(self, name: str):
+  def safe_recv_var(self, name: str):
     """Perform a protected send receive to MAD, by first enabling error handling, so that if an error occurs, an error is received"""
     return self.send(f"{self.py_name}:__err(true):send({name}):__err(false)").recv(name)
 
-  def errhdlr(self, on_off: bool):
+  def set_error_handler(self, on_off: bool):
     """Enable or disable error handling"""
     self.send(f"{self.py_name}:__err({str(on_off).lower()})")
 
   def recv(self, varname: str = None):
     """Receive data from MAD, if a function is returned, it will be executed with the argument mad_communication"""
-    typ = self.ffrom_mad.read(4).decode("utf-8")
+    typ = self.mad_read_stream.read(4).decode("utf-8")
     self.varname = varname  # For mad reference
     return type_fun[typ]["recv"](self)
 
@@ -119,7 +148,7 @@ class mad_process:
      # Check if user has already defined mad (madp_object will have mad defined), otherwise define it
     try:             env["mad"] 
     except KeyError: env["mad"] = self
-    exec(compile(self.recv(), "ffrom_mad", "exec"), self.globalVars, env)
+    exec(compile(self.recv(), "ffrom_mad", "exec"), self.python_exec_context, env)
     return env
 
   # ----------------- Dealing with communication of variables ---------------- #
@@ -133,17 +162,24 @@ class mad_process:
   def recv_vars(self, *names):
     if len(names) == 1:
       if not is_private(names[0]):
-        return self.precv(names[0])
+        return self.safe_recv_var(names[0])
     else:
-      return tuple(self.precv(name) for name in names if not is_private(name))
+      return tuple(self.safe_recv_var(name) for name in names if not is_private(name))
 
   # -------------------------------------------------------------------------- #
 
   def __del__(self):
-    self.send(f"{self.py_name}:__fin()")
-    self.ffrom_mad.close()
-    self.process.terminate()  # In case user left mad waiting
-    self.fto_mad.close()
+    if self.process.poll() is None: # If process is still running
+      self.send(f"{self.py_name}:__fin()") # Tell the mad side to finish
+      self.process.terminate() # Terminate the process on the python side
+
+    # Close the pipes
+    if not self.mad_read_stream.closed:  
+      self.mad_read_stream.close()
+    if not self.mad_input_stream.closed: 
+      self.mad_input_stream.close()
+    
+    # Wait for the process to finish
     self.process.wait()
 
 
@@ -163,11 +199,11 @@ class mad_ref(object):
 
   def __getitem__(self, item: Union[str, int]):
     if isinstance(item, int):
-      result = self._mad.precv(f"{self._name}[{item+1}]")
+      result = self._mad.safe_recv_var(f"{self._name}[{item+1}]")
       if result is None:
         raise IndexError(item)  # For python
     elif isinstance(item, str):
-      result = self._mad.precv(f"{self._name}['{item}']")
+      result = self._mad.safe_recv_var(f"{self._name}['{item}']")
       if result is None:
         raise KeyError(item)  # For python
     else:
@@ -183,12 +219,12 @@ class mad_ref(object):
 # Data ----------------------------------------------------------------------- #
 
 
-def send_dat(self: mad_process, dat_fmt: str, *dat: Any):
-  self.fto_mad.write(struct.pack(dat_fmt, *dat))
+def write_serial_data(self: mad_process, dat_fmt: str, *dat: Any):
+  self.mad_input_stream.write(struct.pack(dat_fmt, *dat))
 
 
-def recv_dat(self: mad_process, dat_sz: int, dat_typ: np.dtype):
-  return np.frombuffer(self.ffrom_mad.read(dat_sz), dtype=dat_typ)
+def read_data_stream(self: mad_process, dat_sz: int, dat_typ: np.dtype):
+  return np.frombuffer(self.mad_read_stream.read(dat_sz), dtype=dat_typ)
 
 
 # None ----------------------------------------------------------------------- #
@@ -201,108 +237,109 @@ def recv_nil(self: mad_process):
 
 # Boolean -------------------------------------------------------------------- #
 def send_bool(self: mad_process, input: bool):
-  return self.fto_mad.write(struct.pack("?", input))
+  return self.mad_input_stream.write(struct.pack("?", input))
 
 def recv_bool(self: mad_process) -> bool:
-  return recv_dat(self, 1, np.bool_)[0]
+  return read_data_stream(self, 1, np.bool_)[0]
 
 # int32 ---------------------------------------------------------------------- #
 
 def send_int(self: mad_process, input: int):
-  return send_dat(self, "i", input)
+  return write_serial_data(self, "i", input)
 
 def recv_int(self: mad_process) -> int:
-  return recv_dat(self, 4, np.int32)[0]
+  return read_data_stream(self, 4, np.int32)[0]
 
 # String --------------------------------------------------------------------- #
 
 
 def send_str(self: mad_process, input: str):
+  self.history += input + "\n"
   send_int(self, len(input))
-  self.fto_mad.write(input.encode("utf-8"))
+  self.mad_input_stream.write(input.encode("utf-8"))
 
 
 def recv_str(self: mad_process) -> str:
-  res = self.ffrom_mad.read(recv_int(self)).decode("utf-8")
+  res = self.mad_read_stream.read(recv_int(self)).decode("utf-8")
   return res
 
 
-# number (float64) ----------------------------------------------------------- #
+# number (in lua, float64 in python) ----------------------------------------- #
 
 def send_num(self: mad_process, input: float):
-  return send_dat(self, "d", input)
+  return write_serial_data(self, "d", input)
 
 def recv_num(self: mad_process) -> float:
-  return recv_dat(self, 8, np.float64)[0]
+  return read_data_stream(self, 8, np.float64)[0]
 
 # Complex (complex128) ------------------------------------------------------- #
 
 def send_cpx(self: mad_process, input: complex):
-  return send_dat(self, "dd", input.real, input.imag)
+  return write_serial_data(self, "dd", input.real, input.imag)
 
 def recv_cpx(self: mad_process) -> complex:
-  return recv_dat(self, 16, np.complex128)[0]
+  return read_data_stream(self, 16, np.complex128)[0]
 
 # Range ---------------------------------------------------------------------- #
 
-def send_grng(self, start: float, stop: float, size: int):
-  send_dat(self, "ddi", start, stop, size)
+def send_generic_range(self, start: float, stop: float, size: int):
+  write_serial_data(self, "ddi", start, stop, size)
 
-def recv_rng(self: mad_process) -> np.ndarray:
-  return np.linspace(*struct.unpack("ddi", self.ffrom_mad.read(20)))
+def recv_range(self: mad_process) -> np.ndarray:
+  return np.linspace(*struct.unpack("ddi", self.mad_read_stream.read(20)))
 
 
-def recv_lrng(self: mad_process) -> np.ndarray:
-  return np.geomspace(*struct.unpack("ddi", self.ffrom_mad.read(20)))
+def recv_logrange(self: mad_process) -> np.ndarray:
+  return np.geomspace(*struct.unpack("ddi", self.mad_read_stream.read(20)))
 
 
 # irange --------------------------------------------------------------------- #
 
-def send_irng(self: mad_process, rng: range):
-  return send_dat(self, "iii", rng.start, rng.stop, rng.step)
+def send_int_range(self: mad_process, rng: range):
+  return write_serial_data(self, "iii", rng.start, rng.stop, rng.step)
 
-def recv_irng(self: mad_process) -> range:
-  start, stop, step = recv_dat(self, 12, np.int32)
+def recv_int_range(self: mad_process) -> range:
+  start, stop, step = read_data_stream(self, 12, np.int32)
   return range(start, stop + 1, step)  # MAD is inclusive at both ends
 
 
 # matrix --------------------------------------------------------------------- #
 
 
-def send_gmat(self: mad_process, mat: np.ndarray):
+def send_generic_matrix(self: mad_process, mat: np.ndarray):
   assert len(mat.shape) == 2, "Matrix must be of two dimensions"
-  send_dat(self, "ii", *mat.shape)
-  self.fto_mad.write(mat.tobytes())
+  write_serial_data(self, "ii", *mat.shape)
+  self.mad_input_stream.write(mat.tobytes())
 
 
-def recv_gmat(self: mad_process, dtype: np.dtype) -> str:
-  shape = recv_dat(self, 8, np.int32)
-  return recv_dat(self, shape[0] * shape[1] * dtype.itemsize, dtype).reshape(shape)
+def recv_generic_matrix(self: mad_process, dtype: np.dtype) -> str:
+  shape = read_data_stream(self, 8, np.int32)
+  return read_data_stream(self, shape[0] * shape[1] * dtype.itemsize, dtype).reshape(shape)
 
 
-def recv_mat(self: mad_process) -> np.ndarray:
-  return recv_gmat(self, np.dtype("float64"))
+def recv_matrix(self: mad_process) -> np.ndarray:
+  return recv_generic_matrix(self, np.dtype("float64"))
                    
-def recv_cmat(self: mad_process) -> np.ndarray:
-  return recv_gmat(self, np.dtype("complex128"))
+def recv_cpx_matrix(self: mad_process) -> np.ndarray:
+  return recv_generic_matrix(self, np.dtype("complex128"))
                    
-def recv_imat(self: mad_process) -> np.ndarray:
-  return recv_gmat(self, np.dtype("int32"))
+def recv_int_matrix(self: mad_process) -> np.ndarray:
+  return recv_generic_matrix(self, np.dtype("int32"))
 
 # monomial ------------------------------------------------------------------- #
 
 
-def send_mono(self: mad_process, mono: np.ndarray):
+def send_monomial(self: mad_process, mono: np.ndarray):
   send_int(self, mono.size)
-  self.fto_mad.write(mono.tobytes())
+  self.mad_input_stream.write(mono.tobytes())
 
-def recv_mono(self: mad_process) -> np.ndarray:
-  return recv_dat(self, recv_int(self), np.ubyte)
+def recv_monomial(self: mad_process) -> np.ndarray:
+  return read_data_stream(self, recv_int(self), np.ubyte)
 
 # TPSA ----------------------------------------------------------------------- #
 
 
-def send_gtpsa(
+def send_generic_tpsa(
   self: mad_process,
   monos: np.ndarray,
   coefficients: np.ndarray,
@@ -311,35 +348,35 @@ def send_gtpsa(
   assert len(monos.shape) == 2, "The list of monomials must have two dimensions"
   assert len(monos) == len(coefficients), "The number of monomials must be equal to the number of coefficients"
   assert monos.dtype == np.uint8, "The monomials must be of type 8-bit unsigned integer "
-  send_dat(self, "ii", len(monos), len(monos[0]))
+  write_serial_data(self, "ii", len(monos), len(monos[0]))
   for mono in monos:
-    self.fto_mad.write(mono.tobytes())
+    self.mad_input_stream.write(mono.tobytes())
   for coefficient in coefficients:
     send_num(self, coefficient)
 
 
-def recv_gtpsa(self: mad_process, dtype: np.dtype) -> np.ndarray:
-  num_mono, mono_len = recv_dat(self, 8, np.int32)
+def recv_generic_tpsa(self: mad_process, dtype: np.dtype) -> np.ndarray:
+  num_mono, mono_len = read_data_stream(self, 8, np.int32)
   mono_list = np.reshape(
-    recv_dat(self, mono_len * num_mono, np.ubyte),
+    read_data_stream(self, mono_len * num_mono, np.ubyte),
     (num_mono, mono_len),
   )
-  coefficients = recv_dat(self, num_mono * dtype.itemsize, dtype)
+  coefficients = read_data_stream(self, num_mono * dtype.itemsize, dtype)
   return mono_list, coefficients
 
 
-def recv_ctpa(self: mad_process):
-  return recv_gtpsa(self, np.dtype("complex128"))
+def recv_cpx_tpsa(self: mad_process):
+  return recv_generic_tpsa(self, np.dtype("complex128"))
 
-def recv_tpsa(self: mad_process):
-  return recv_gtpsa(self, np.dtype("float64"))
+def recv_dbl_tpsa(self: mad_process):
+  return recv_generic_tpsa(self, np.dtype("float64"))
 
-recv_ctpa = lambda self: recv_gtpsa(self, np.dtype("complex128"))
-recv_tpsa = lambda self: recv_gtpsa(self, np.dtype("float64"))
+recv_cpx_tpsa = lambda self: recv_generic_tpsa(self, np.dtype("complex128"))
+recv_dbl_tpsa = lambda self: recv_generic_tpsa(self, np.dtype("float64"))
 
 # lists ---------------------------------------------------------------------- #
 
-
+# Lists of strings are really slow to send, is there a way to improve this? (jgray 2024)
 def send_list(self: mad_process, lst: list):
   send_int(self, len(lst))
   for item in lst:
@@ -360,42 +397,42 @@ def recv_list(self: mad_process) -> list:
 
 # object (table with metatable are treated as pure reference) ---------------- #
 
-def recv_ref(self: mad_process):
+def recv_reference(self: mad_process):
   return mad_ref(self.varname, self)
 
-def send_ref(self, obj: mad_ref):
+def send_reference(self, obj: mad_ref):
   return send_str(self, f"return {obj._name}")
 
 # error ---------------------------------------------------------------------- #
 
 
 def recv_err(self: mad_process):
-  self.errhdlr(False)
+  self.set_error_handler(False)
   raise RuntimeError("MAD Errored (see the MAD error output)")
 
 
 # ---------------------------- dispatch tables ------------------------------- #
 type_fun = {
-  "nil_": {"recv": recv_nil , "send": send_nil },
-  "bool": {"recv": recv_bool, "send": send_bool},
-  "str_": {"recv": recv_str , "send": send_str },
-  "tbl_": {"recv": recv_list, "send": send_list},
-  "ref_": {"recv": recv_ref , "send": send_ref },
-  "fun_": {"recv": recv_ref , "send": send_ref },
-  "obj_": {"recv": recv_ref , "send": send_ref },
-  "int_": {"recv": recv_int , "send": send_int },
-  "num_": {"recv": recv_num , "send": send_num },
-  "cpx_": {"recv": recv_cpx , "send": send_cpx },
-  "mat_": {"recv": recv_mat , "send": send_gmat},
-  "cmat": {"recv": recv_cmat, "send": send_gmat},
-  "imat": {"recv": recv_imat, "send": send_gmat},
-  "rng_": {"recv": recv_rng ,                  },
-  "lrng": {"recv": recv_lrng,                  },
-  "irng": {"recv": recv_irng, "send": send_irng},
-  "mono": {"recv": recv_mono, "send": send_mono},
-  "tpsa": {"recv": recv_tpsa,                  },
-  "ctpa": {"recv": recv_ctpa,                  },
-  "err_": {"recv": recv_err ,                  },
+  "nil_": {"recv": recv_nil       , "send": send_nil           },
+  "bool": {"recv": recv_bool      , "send": send_bool          },
+  "str_": {"recv": recv_str       , "send": send_str           },
+  "tbl_": {"recv": recv_list      , "send": send_list          },
+  "ref_": {"recv": recv_reference , "send": send_reference     },
+  "fun_": {"recv": recv_reference , "send": send_reference     },
+  "obj_": {"recv": recv_reference , "send": send_reference     },
+  "int_": {"recv": recv_int       , "send": send_int           },
+  "num_": {"recv": recv_num       , "send": send_num           },
+  "cpx_": {"recv": recv_cpx       , "send": send_cpx           },
+  "mat_": {"recv": recv_matrix    , "send": send_generic_matrix},
+  "cmat": {"recv": recv_cpx_matrix, "send": send_generic_matrix},
+  "imat": {"recv": recv_int_matrix, "send": send_generic_matrix},
+  "rng_": {"recv": recv_range     ,                            },
+  "lrng": {"recv": recv_logrange  ,                            },
+  "irng": {"recv": recv_int_range , "send": send_int_range     },
+  "mono": {"recv": recv_monomial  , "send": send_monomial      },
+  "tpsa": {"recv": recv_dbl_tpsa,                              },
+  "ctpa": {"recv": recv_cpx_tpsa,                              },
+  "err_": {"recv": recv_err ,                                  },
 }